Refactoring stream reader.
