+++ I3DShapesTool/I3DShapesHeader.cs
@@ -1,8 +1,40 @@
-﻿namespace I3DShapesTool
+﻿using System;
+using System.IO;
+
+namespace I3DShapesTool
 {
-    struct I3DShapesHeader
+    public class I3DShapesHeader
     {
-        public short Version;
-        public byte Seed;
+        public short Version { get; }
+        public byte Seed { get; }
+
+        public I3DShapesHeader(Stream stream)
+        {
+            var b1 = (byte)stream.ReadByte();
+            var b2 = (byte)stream.ReadByte();
+            var b3 = (byte)stream.ReadByte();
+            var b4 = (byte)stream.ReadByte();
+
+            byte seed;
+            short version;
+
+            if (b1 >= 4) // Might be 5 as well
+            {
+                version = b1;
+                seed = b3;
+            }
+            else if (b4 == 2 || b4 == 3)
+            {
+                version = b4;
+                seed = b2;
+            }
+            else
+            {
+                throw new NotSupportedException("Unknown version");
+            }
+
+            Seed = seed;
+            Version = version;
+        }
     }
 }
